+++ frontend/src/features/layersSelector/search/slice.ts
