+++ frontend/src/hooks/useAppSelector.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-restricted-imports */
 import { useSelector } from 'react-redux'
 
 import type { HomeRootState } from '../store'
